update namespacing for converters
