+++ src/electionguard-ui/ElectionGuard.UI/Controls/ContestControl.xaml
@@ -3,13 +3,14 @@
     x:Class="ElectionGuard.UI.Controls.ContestControl"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:converter="clr-namespace:ElectionGuard.UI.Converters"
     xmlns:helper="clr-namespace:ElectionGuard.UI.Helpers"
     xmlns:model="clr-namespace:ElectionGuard.UI.Lib.Models;assembly=ElectionGuard.UI.Lib"
     x:Name="this"
     x:DataType="model:ContestItem">
     <ContentView.Resources>
         <ResourceDictionary>
-            <helper:AlternateRowColorConverter x:Key="AlternateRowColorConverter" />
+            <converter:AlternateRowColorConverter x:Key="AlternateRowColorConverter" />
             <Style TargetType="Label">
                 <Setter Property="VerticalOptions" Value="Start" />
             </Style>
