+++ src/electionguard-ui/ElectionGuard.UI/Converters/BackupConverter.cs
@@ -1,0 +1,16 @@
+﻿using System.Globalization;
+
+namespace ElectionGuard.UI.Converters;
+
+internal class BackupConverter : IValueConverter
+{
+    public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+    {
+        return (bool)value ? "backup.png" : "notbackup.png";
+    }
+
+    public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+    {
+        throw new NotImplementedException();
+    }
+}
