+++ src/electionguard-ui/ElectionGuard.UI/Converters/GuardianBackgroundConverter.cs
@@ -1,0 +1,16 @@
+﻿using System.Globalization;
+
+namespace ElectionGuard.UI.Converters;
+
+class GuardianBackgroundConverter : IValueConverter
+{
+    public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+    {
+        return (bool)value ? Colors.LightGrey : Colors.Transparent;
+    }
+
+    public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+    {
+        throw new NotImplementedException();
+    }
+}
