--- src/electionguard-ui/ElectionGuard.UI/Helpers/BackupConverter.cs
@@ -1,16 +1,0 @@
-﻿using System.Globalization;
-
-namespace ElectionGuard.UI.Helpers;
-
-internal class BackupConverter : IValueConverter
-{
-    public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-    {
-        return (bool)value ? "backup.png" : "notbackup.png";
-    }
-
-    public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-    {
-        throw new NotImplementedException();
-    }
-}
