--- src/electionguard-ui/ElectionGuard.UI/Helpers/TallyCanBeAbandonedConverter.cs
@@ -1,16 +1,0 @@
-﻿using System.Globalization;
-using CommunityToolkit.Maui.Converters;
-
-namespace ElectionGuard.UI.Helpers
-{
-    internal class TallyCanBeAbandonedConverter
-        : BaseConverter<TallyState, bool>
-    {
-        public override bool DefaultConvertReturnValue { get; set; }
-        public override TallyState DefaultConvertBackReturnValue { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-
-        public override TallyState ConvertBackTo(bool value, CultureInfo? culture) => throw new NotImplementedException();
-
-        public override bool ConvertFrom(TallyState value, CultureInfo? culture) => value < TallyState.TallyStarted;
-    }
-}
