+++ src/electionguard-ui/ElectionGuard.UI/Views/BallotUploadPage.xaml
@@ -3,6 +3,7 @@
     x:Class="ElectionGuard.UI.Views.BallotUploadPage"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:converter="clr-namespace:ElectionGuard.UI.Converters"
     xmlns:helper="clr-namespace:ElectionGuard.UI.Helpers"
     xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
     xmlns:viewModels="clr-namespace:ElectionGuard.UI.ViewModels"
@@ -12,8 +13,8 @@
     x:TypeArguments="viewModels:BallotUploadViewModel"
     ControlTemplate="{StaticResource PageTemplate}">
     <ContentPage.Resources>
-        <helper:BallotUloadPanelConverter x:Key="BallotUloadPanelConverter" />
-        <helper:IsNotEmptyConverter x:Key="IsNotEmptyConverter" />
+        <converter:BallotUloadPanelConverter x:Key="BallotUloadPanelConverter" />
+        <converter:IsNotEmptyConverter x:Key="IsNotEmptyConverter" />
     </ContentPage.Resources>
     <Grid Style="{StaticResource NarrowCenteredGrid}">
         <VerticalStackLayout IsVisible="{Binding ShowPanel, Converter={StaticResource BallotUloadPanelConverter}, ConverterParameter=AutoUpload}">
