+++ src/electionguard-ui/ElectionGuard.UI/Views/ElectionPage.xaml
@@ -4,6 +4,7 @@
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:controls="clr-namespace:ElectionGuard.UI.Controls"
+    xmlns:converter="clr-namespace:ElectionGuard.UI.Converters"
     xmlns:helper="clr-namespace:ElectionGuard.UI.Helpers"
     xmlns:model="clr-namespace:ElectionGuard.UI.Lib.Models;assembly=ElectionGuard.UI.Lib"
     xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
@@ -14,7 +15,7 @@
     ControlTemplate="{StaticResource PageTemplate}">
     <ContentPage.Resources>
         <ResourceDictionary>
-            <helper:AlternateRowColorConverter x:Key="AlternateRowColorConverter" />
+            <converter:AlternateRowColorConverter x:Key="AlternateRowColorConverter" />
         </ResourceDictionary>
     </ContentPage.Resources>
     <Grid
