+++ src/electionguard-ui/ElectionGuard.UI/Views/ViewKeyCeremonyPage.xaml
@@ -4,6 +4,7 @@
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:controls="clr-namespace:ElectionGuard.UI.Controls"
+    xmlns:converter="clr-namespace:ElectionGuard.UI.Converters"
     xmlns:helper="clr-namespace:ElectionGuard.UI.Helpers"
     xmlns:model="clr-namespace:ElectionGuard.UI.Lib.Models;assembly=ElectionGuard.UI.Lib"
     xmlns:models="clr-namespace:ElectionGuard.UI.Models"
@@ -20,11 +21,11 @@
                 x:Key="BoolToObjectConverter"
                 FalseObject="false"
                 TrueObject="true" />
-            <helper:StateConverter x:Key="StateConverter" />
-            <helper:BackupConverter x:Key="BackupConverter" />
-            <helper:BadVerifyConverter x:Key="BadVerifyConverter" />
-            <helper:BadVerifyTooltipConverter x:Key="BadVerifyTooltipConverter" />
-            <helper:GuardianBackgroundConverter x:Key="GuardianBackgroundConverter" />
+            <converter:StateConverter x:Key="StateConverter" />
+            <converter:BackupConverter x:Key="BackupConverter" />
+            <converter:BadVerifyConverter x:Key="BadVerifyConverter" />
+            <converter:BadVerifyTooltipConverter x:Key="BadVerifyTooltipConverter" />
+            <converter:GuardianBackgroundConverter x:Key="GuardianBackgroundConverter" />
         </ResourceDictionary>
     </ContentPage.Resources>
